Undo monitoring stuff, we only want to fix the Frank!Doc
