+++ core/src/main/java/nl/nn/adapterframework/monitoring/MonitorManager.java
@@ -34,7 +34,6 @@ import lombok.Getter;
 import lombok.Setter;
 import nl.nn.adapterframework.configuration.ConfigurationException;
 import nl.nn.adapterframework.core.Adapter;
-import nl.nn.adapterframework.doc.FrankDocGroup;
 import nl.nn.adapterframework.lifecycle.ConfigurableLifecyleBase;
 import nl.nn.adapterframework.monitoring.events.Event;
 import nl.nn.adapterframework.monitoring.events.RegisterMonitorEvent;
@@ -51,7 +50,6 @@ import nl.nn.adapterframework.util.XmlBuilder;
  * @author Niels Meijer
  * @version 2.0
  */
-@FrankDocGroup(name = "Monitoring")
 public class MonitorManager extends ConfigurableLifecyleBase implements ApplicationContextAware, ApplicationListener<RegisterMonitorEvent> {
 
 	private @Getter @Setter ApplicationContext applicationContext;
